[dap] Don't use --start-paused when running in Profile/Release mode (#98926)
