+++ packages/flutter_tools/test/general.shard/dap/flutter_test_adapter_test.dart
@@ -12,15 +12,24 @@ import 'package:test/test.dart';
 import 'mocks.dart';
 
 void main() {
-  group('flutter adapter', () {
+  group('flutter test adapter', () {
+    final String expectedFlutterExecutable = globals.platform.isWindows
+        ? r'C:\fake\flutter\bin\flutter.bat'
+        : '/fake/flutter/bin/flutter';
+
     setUpAll(() {
-    Cache.flutterRoot = '/fake/flutter';
-  });
+      Cache.flutterRoot = globals.platform.isWindows
+          ? r'C:\fake\flutter'
+          : '/fake/flutter';
+    });
 
     test('includes toolArgs', () async {
-      final MockFlutterDebugAdapter adapter = MockFlutterDebugAdapter(fileSystem: globals.fs, platform: globals.platform);
+      final MockFlutterTestDebugAdapter adapter = MockFlutterTestDebugAdapter(
+        fileSystem: globals.fs,
+        platform: globals.platform,
+      );
       final Completer<void> responseCompleter = Completer<void>();
-
+      final MockRequest request = MockRequest();
       final FlutterLaunchRequestArguments args = FlutterLaunchRequestArguments(
         cwd: '/project',
         program: 'foo.dart',
@@ -28,17 +37,20 @@ void main() {
         noDebug: true,
       );
 
-      await adapter.configurationDoneRequest(MockRequest(), null, () {});
-      await adapter.launchRequest(MockRequest(), args, responseCompleter.complete);
+      await adapter.configurationDoneRequest(request, null, () {});
+      await adapter.launchRequest(request, args, responseCompleter.complete);
       await responseCompleter.future;
 
-      expect(adapter.executable, equals('/fake/flutter/bin/flutter'));
+      expect(adapter.executable, equals(expectedFlutterExecutable));
       expect(adapter.processArgs, contains('tool_arg'));
     });
 
     group('includes customTool', () {
       test('with no args replaced', () async {
-        final MockFlutterDebugAdapter adapter = MockFlutterDebugAdapter(fileSystem: globals.fs, platform: globals.platform);
+        final MockFlutterTestDebugAdapter adapter = MockFlutterTestDebugAdapter(fileSystem: globals.fs,
+        platform: globals.platform,);
+        final Completer<void> responseCompleter = Completer<void>();
+        final MockRequest request = MockRequest();
         final FlutterLaunchRequestArguments args = FlutterLaunchRequestArguments(
           cwd: '/project',
           program: 'foo.dart',
@@ -46,9 +58,8 @@ void main() {
           noDebug: true,
         );
 
-        await adapter.configurationDoneRequest(MockRequest(), null, () {});
-        final Completer<void> responseCompleter = Completer<void>();
-        await adapter.launchRequest(MockRequest(), args, responseCompleter.complete);
+        await adapter.configurationDoneRequest(request, null, () {});
+        await adapter.launchRequest(request, args, responseCompleter.complete);
         await responseCompleter.future;
 
         expect(adapter.executable, equals('/custom/flutter'));
@@ -57,7 +68,10 @@ void main() {
       });
 
       test('with all args replaced', () async {
-        final MockFlutterDebugAdapter adapter = MockFlutterDebugAdapter(fileSystem: globals.fs, platform: globals.platform);
+        final MockFlutterTestDebugAdapter adapter = MockFlutterTestDebugAdapter(fileSystem: globals.fs,
+        platform: globals.platform,);
+        final Completer<void> responseCompleter = Completer<void>();
+        final MockRequest request = MockRequest();
         final FlutterLaunchRequestArguments args = FlutterLaunchRequestArguments(
           cwd: '/project',
           program: 'foo.dart',
@@ -67,9 +81,8 @@ void main() {
           toolArgs: <String>['tool_args'], // should still be in args
         );
 
-        await adapter.configurationDoneRequest(MockRequest(), null, () {});
-        final Completer<void> responseCompleter = Completer<void>();
-        await adapter.launchRequest(MockRequest(), args, responseCompleter.complete);
+        await adapter.configurationDoneRequest(request, null, () {});
+        await adapter.launchRequest(request, args, responseCompleter.complete);
         await responseCompleter.future;
 
         expect(adapter.executable, equals('/custom/flutter'));
